refactor: 로또 생성 및 파싱 관련 클래스 패키지 구조 개선
