+++ src/main/java/lotto/domain/AutoBasedLottoGenerator.java
@@ -1,12 +1,9 @@
-package lotto.utils;
+package lotto.domain;
 
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import lotto.domain.Lotto;
-import lotto.domain.LottoCount;
-import lotto.domain.LottoNumber;
 
 public class AutoBasedLottoGenerator {
     private static final int LOTTO_NUMBER_MIN = 1;
