+++ src/main/java/lotto/domain/ManualBasedLottoGenerator.java
@@ -1,7 +1,6 @@
-package lotto.utils;
+package lotto.domain;
 
 import java.util.List;
-import lotto.domain.Lotto;
 
 public class ManualBasedLottoGenerator {
