+++ src/test/java/lotto/utils/ManualBasedLottoGeneratorTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.*;
 
 import java.util.List;
 import lotto.domain.Lotto;
+import lotto.domain.ManualBasedLottoGenerator;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
